+++ .env
@@ -1,0 +1,1 @@
+API_URL = "https://abarrotes.msalazar.dev/jsonapi/node/"
