+++ babel.config.js
@@ -2,6 +2,7 @@ module.exports = function(api) {
   api.cache(true);
   return {
     presets: ['babel-preset-expo'],
-    plugins: ['react-native-reanimated/plugin']
+    plugins: ['react-native-reanimated/plugin'],
+    plugins: ['module:react-native-dotenv']
   };
 };
