+++ package.json
@@ -26,6 +26,7 @@
     "react": "18.2.0",
     "react-native": "0.71.7",
     "react-native-animatable": "^1.3.3",
+    "react-native-dotenv": "^3.4.9",
     "react-native-gesture-handler": "~2.9.0",
     "react-native-paper": "^5.5.1",
     "react-native-reanimated": "~2.14.4",
